+++ src/Management/BugInfo.ts
@@ -1,6 +1,6 @@
+import { IBugInfo } from "./IBugInfo.js";
 import { JSONObject } from "../Utilities/JSONObject.js";
 import { JSONObjectBase } from "../Utilities/JSONObjectBase.js";
-import { IBugInfo } from "./IBugInfo.js";
 
 /**
  * Provides info for reporting bugs.