+++ src/Management/PackageDependencyCollection.ts
@@ -1,8 +1,8 @@
-import { Dictionary } from "../Collections/Dictionary.js";
-import { Package } from "../Package.js";
 import { DependencyCollection } from "./DependencyCollection.js";
 import { KeyOfType } from "./KeyOfType.js";
 import { PackageDependencyCollectionOptions } from "./PackageDependencyCollectionOptions.js";
+import { Dictionary } from "../Collections/Dictionary.js";
+import { Package } from "../Package.js";
 
 /**
  * Provides a set of dependencies which are loaded from a package.
@@ -86,31 +86,6 @@ export class PackageDependencyCollection extends DependencyCollection
         return this.LoadDependencies(
             nameof<PackageDependencyCollectionOptions>((o) => o.optionalDependencies) as keyof PackageDependencyCollectionOptions);
     }
-
-    /**
-     * Loads the dependencies specified in the list with the specified {@link listName `listName`} from the {@link PackageDependencyCollection.Package `Package`}.
-     *
-     * @param listName
-     * The name of the list of dependency-names to load.
-     *
-     * @returns
-     * The loaded dependencies.
-     */
-    protected LoadDependencies(listName: keyof PackageDependencyCollectionOptions): Dictionary<string, string>;
-
-    /**
-     * Loads the dependencies specified in the list with the specified {@link listName `listName`} from the specified {@link packageListName `packageListName`} in the {@link PackageDependencyCollection.Package `Package`}.
-     *
-     * @param listName
-     * The name of the list of dependency-names to load.
-     *
-     * @param packageListName
-     * The name of the list of the package to load the dependency-versions from.
-     *
-     * @returns
-     * The loaded dependencies.
-     */
-    protected LoadDependencies(listName: keyof PackageDependencyCollectionOptions, packageListName: KeyOfType<DependencyCollection, Dictionary<string, string>>): Dictionary<string, string>;
 
     /**
      * Loads the dependencies specified in the list with the specified {@link listName `listName`} from the specified {@link packageListName `packageListName`} in the {@link PackageDependencyCollection.Package `Package`}.