+++ CHANGELOG.md
@@ -6,7 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## NPMPackageEditor [Unreleased]
 
-[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v3.1.0...dev)
+[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v3.1.1...dev)
+
+## NPMPackageEditor v3.1.1
+### Fixed
+  - Error that caused the `private` property to be added to dumped packages even if it is set to `false`
+
+### Added
+  - Support for dumping properties only if they are set to a truthy value by introducing `GenerationLogic.Truthy`
+
+### Updated
+  - All dependencies
+
+[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v3.1.0...v3.1.1)
 
 ## NPMPackageEditor v3.1.0
 ### Fixed